app/requirements.txt
ccxt
PyMysql
Flask
gunicorn
flask_sqlalchemy
passlib
flask_jwt_extended
flask_restful
python-dotenv
flask_orator